+++ internal/provider/kubernetes/kubernetes.go
@@ -50,11 +50,8 @@ func New(cfg *rest.Config, svr *config.Server, resources *message.ProviderResour
 		return nil, fmt.Errorf("failed to add status update handler %v", err)
 	}
 
-	// Initialize kubernetes provider referenceStore to store additional object mappings.
-	referenceStore := newProviderReferenceStore()
-
 	// Create and register the controllers with the manager.
-	if err := newGatewayAPIController(mgr, svr, updateHandler.Writer(), resources, referenceStore); err != nil {
+	if err := newGatewayAPIController(mgr, svr, updateHandler.Writer(), resources); err != nil {
 		return nil, fmt.Errorf("failted to create gatewayapi controller: %w", err)
 	}
