+++ internal/provider/kubernetes/routes.go
@@ -1,0 +1,120 @@
+// Copyright Envoy Gateway Authors
+// SPDX-License-Identifier: Apache-2.0
+// The full text of the Apache license is available in the LICENSE file at
+// the root of the repo.
+
+package kubernetes
+
+import (
+	"context"
+
+	"github.com/envoyproxy/gateway/internal/gatewayapi"
+	"github.com/envoyproxy/gateway/internal/provider/utils"
+	"k8s.io/apimachinery/pkg/fields"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	gwapiv1a2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
+	gwapiv1b1 "sigs.k8s.io/gateway-api/apis/v1beta1"
+)
+
+// processTLSRoutes finds TLSRoutes corresponding to a gatewayNamespaceName, further checks for
+// the backend references and pushes the TLSRoutes to the resourceTree.
+func (r *gatewayAPIReconciler) processTLSRoutes(ctx context.Context, gatewayNamespaceName string,
+	resourceMap *resourceMappings, resourceTree *gatewayapi.Resources) error {
+	tlsRouteList := &gwapiv1a2.TLSRouteList{}
+	if err := r.client.List(ctx, tlsRouteList, &client.ListOptions{
+		FieldSelector: fields.OneTermEqualSelector(gatewayTLSRouteIndex, gatewayNamespaceName),
+	}); err != nil {
+		r.log.Error(err, "unable to find associated TLSRoutes")
+		return err
+	}
+
+	for _, tlsRoute := range tlsRouteList.Items {
+		tlsRoute := tlsRoute
+		r.log.Info("processing TLSRoute", "namespace", tlsRoute.Namespace, "name", tlsRoute.Name)
+
+		for _, rule := range tlsRoute.Spec.Rules {
+			for _, backendRef := range rule.BackendRefs {
+				backendRef := backendRef
+				ref := gatewayapi.UpgradeBackendRef(backendRef)
+				if err := validateBackendRef(&ref); err != nil {
+					r.log.Error(err, "invalid backendRef")
+					continue
+				}
+
+				backendNamespace := gatewayapi.NamespaceDerefOrAlpha(backendRef.Namespace, tlsRoute.Namespace)
+				resourceMap.allAssociatedBackendRefs[types.NamespacedName{
+					Namespace: backendNamespace,
+					Name:      string(backendRef.Name),
+				}] = struct{}{}
+
+				if backendNamespace != tlsRoute.Namespace {
+					from := ObjectKindNamespacedName{kind: gatewayapi.KindTLSRoute, namespace: tlsRoute.Namespace, name: tlsRoute.Name}
+					to := ObjectKindNamespacedName{kind: gatewayapi.KindService, namespace: backendNamespace, name: string(backendRef.Name)}
+					refGrant, err := r.findReferenceGrant(ctx, from, to)
+					if err != nil {
+						r.log.Error(err, "unable to find ReferenceGrant that links the Service to TLSRoute")
+						continue
+					}
+
+					resourceMap.allAssociatedRefGrants[utils.NamespacedName(refGrant)] = refGrant
+				}
+			}
+		}
+
+		resourceMap.allAssociatedNamespaces[tlsRoute.Namespace] = struct{}{}
+		resourceTree.TLSRoutes = append(resourceTree.TLSRoutes, &tlsRoute)
+	}
+
+	return nil
+}
+
+// processHTTPRoutes finds HTTPRoutes corresponding to a gatewayNamespaceName, further checks for
+// the backend references and pushes the HTTPRoutes to the resourceTree.
+func (r *gatewayAPIReconciler) processHTTPRoutes(ctx context.Context, gatewayNamespaceName string,
+	resourceMap *resourceMappings, resourceTree *gatewayapi.Resources) error {
+	httpRouteList := &gwapiv1b1.HTTPRouteList{}
+	if err := r.client.List(ctx, httpRouteList, &client.ListOptions{
+		FieldSelector: fields.OneTermEqualSelector(gatewayHTTPRouteIndex, gatewayNamespaceName),
+	}); err != nil {
+		r.log.Error(err, "unable to find associated HTTPRoutes")
+		return err
+	}
+	for _, httpRoute := range httpRouteList.Items {
+		httpRoute := httpRoute
+		r.log.Info("processing HTTPRoute", "namespace", httpRoute.Namespace, "name", httpRoute.Name)
+
+		for _, rule := range httpRoute.Spec.Rules {
+			for _, backendRef := range rule.BackendRefs {
+				backendRef := backendRef
+				if err := validateBackendRef(&backendRef.BackendRef); err != nil {
+					r.log.Error(err, "invalid backendRef")
+					continue
+				}
+
+				backendNamespace := gatewayapi.NamespaceDerefOr(backendRef.Namespace, httpRoute.Namespace)
+				resourceMap.allAssociatedBackendRefs[types.NamespacedName{
+					Namespace: backendNamespace,
+					Name:      string(backendRef.Name),
+				}] = struct{}{}
+
+				if backendNamespace != httpRoute.Namespace {
+					from := ObjectKindNamespacedName{kind: gatewayapi.KindHTTPRoute, namespace: httpRoute.Namespace, name: httpRoute.Name}
+					to := ObjectKindNamespacedName{kind: gatewayapi.KindService, namespace: backendNamespace, name: string(backendRef.Name)}
+					refGrant, err := r.findReferenceGrant(ctx, from, to)
+					if err != nil {
+						r.log.Error(err, "unable to find ReferenceGrant that links the Service to HTTPRoute")
+						continue
+					}
+
+					resourceMap.allAssociatedRefGrants[utils.NamespacedName(refGrant)] = refGrant
+				}
+			}
+		}
+
+		resourceMap.allAssociatedNamespaces[httpRoute.Namespace] = struct{}{}
+		resourceTree.HTTPRoutes = append(resourceTree.HTTPRoutes, &httpRoute)
+	}
+
+	return nil
+}
